Feat: HomePage에 Event를 Coupon으로 변경
