+++ app/src/main/java/dev/kichan/marketplace/ui/component/atoms/MyPageCard.kt
@@ -1,35 +1,16 @@
 package dev.kichan.marketplace.ui.component.dev.kichan.marketplace.ui.component.atoms
 
-import Carbon_bookmark
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material.icons.filled.LocationOn
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import dev.kichan.marketplace.R
-import dev.kichan.marketplace.ui.data.Event
+import dev.kichan.marketplace.ui.data.CouponBoxProps
 
 @Composable
-fun MyPageCard(event: Event, modifier: Modifier = Modifier) {
+fun MyPageCard(couponBoxProps: CouponBoxProps, modifier: Modifier = Modifier) {
     var isBookmarked by remember { mutableStateOf(false) }
 
 //    Row(
