+++ app/src/main/java/dev/kichan/marketplace/ui/data/CouponBoxProps.kt
@@ -1,0 +1,11 @@
+package dev.kichan.marketplace.ui.data
+
+data class CouponBoxProps(
+    val id: String,
+    val title: String,
+    val subTitle : String,
+    val url : String,
+    val marketId: Long,
+    val isDownload: Boolean,
+    val onDownloadClick: () -> Unit,
+)
