--- app/src/main/java/dev/kichan/marketplace/ui/data/Event.kt
@@ -1,11 +1,0 @@
-package dev.kichan.marketplace.ui.data
-
-data class Event(
-    val id: String,
-    val title: String,
-    val subTitle : String,
-    val url : String,
-    val marketId: Long,
-    val isDownload: Boolean,
-    val onDownloadClick: () -> Unit,
-)
